pyproject.toml
[project]
name = "llm-as-a-judge"
version = "0.1.0"
description = "LLM as a judge for evaluating attack paths"
authors = [
    { name = "Andrei Moise", email = "neuronspeeed@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.0,<2.0.0",
    "openai>=1.3.0,<2.0.0",
    "langchain>=0.1.0,<0.2.0",
    "langchain-openai>=0.1.0,<0.2.0",
    "pydantic>=2.8.0,<3.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/llm_as_a_judge"]

[tool.rye.scripts]
start = "python src/llm_as_a_judge/main.py"